+++ app/components/AppProduct.vue
@@ -225,7 +225,9 @@ export default {
         this.supports_high_res_granules = this.initial_product.highres_available;
         this.granules = await this.getGranules();
         if (this.granules_selection_mode === 'single') {
-          this.selected_granule_indexes = [this.granules.length - 1];
+          if (this.granules.length < 0) {
+            this.selected_granule_indexes = [this.granules.length - 1];
+          }
         }
       }
     },